+++ src/main/java/com/example/authenpaymentservice/authen/service/AuthService.java
@@ -203,7 +203,7 @@ public class AuthService extends BaseService implements UserDetailsService {
         return new CustomUserDetails(user, authorities);
     }
 
-    public UserDetails loadUserById(int id) {
+    public UserDetails loadUserById(long id) {
         User user = userRepository.findUserById(id);
         List<SimpleGrantedAuthority> authorities = new ArrayList<>();
         authorities.add(new SimpleGrantedAuthority(UserRole.MEMBER.toString()));
@@ -228,7 +228,7 @@ public class AuthService extends BaseService implements UserDetailsService {
         userRepository.saveAndFlush(user);
     }
 
-    public void cacheRefreshToken(int userId, String refreshToken) throws IOException {
+    public void cacheRefreshToken(long userId, String refreshToken) throws IOException {
         String key = CacheKey.genRefreshToken(userId);
         cacheUtils.set(key, refreshToken, jwtTokenProvider.getRefreshExpireTime());
     }
